backend/requirements.txt
cryptography==46.0.3
fastapi==0.124.4
passlib==1.7.4
pydantic==2.12.5
PyJWT==2.10.1
python-dotenv==1.2.1
SQLAlchemy==2.0.45
uvicorn==0.23.2
psycopg2-binary==2.9.7
python-multipart==0.0.6
alembic==1.11.1